requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
decorator==5.2.1
display==1.0.0
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jedi==0.19.2
jiter==0.10.0
Markdown==3.8.2
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ollama==0.5.3
openai==1.102.0
parso==0.8.5
prompt_toolkit==3.0.52
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
rich==14.1.0
sarvamai==0.1.19
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
